Create getCategories method in CompleteQTO bim-component
